DEPLOYMENT/deployment_project/requirements.txt
# Pinned loosely for Docker layer caching; sync with repo Pipfile for local pipenv.
Django>=6.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.4
python-decouple>=3.8
gunicorn>=23.0
psycopg[binary]>=3.2
whitenoise>=6.8
